bundle: allow specifying ref names in --stdin input
The bundle archive format inherently requires specifying reference
names along with the revision tips contained therein, but it doesn't
inherently require those reference names to be "real", that's always
been an artifact of how create_bundle() was first implemented. See [1]
for that implementation.

If you want to create a bundle of a revision tip or revision range(s)
that a reference name doesn't point to you previously needed trickery
such as creating a temporary reference, or e.g. having a temporary
bare repository point to the "real" repository with
GIT_OBJECT_DIRECTORY (or GIT_ALTERNATE_OBJECT_DIRECTORIES etc.), and
creating a throwaway reference just to craft the bundle.

With this change revision/reference name pairs can be specified in the
--stdin input, see the Documentation/* part of this change below for
the exact behavior.

I think the whole use of "refname_to_pending" here is rather nasty,
but could not think of a better alternative to track a
"rev_info->pending" from the outside.

The semantics are rather straightforward though: As we parse lines on
stdin we have the revision.c API parse the revision part, and it
pushes a revision (or none, or two) to "rev_info->pending". Our
"handle_stdin_line" callback then runs, sees how much (if anything)
got added to "rev_info->pending", and extends a string-list
correspondingly such that we end up with a 1=1 positional mapping
between the "rev_info->pending" and our own "refname_to_pending"
string-list. See the commentary in
write_bundle_after_stdin_line_again() for more details.

1. 30415d50ccc (Move bundle specific stuff into bundle.[ch],
   2007-09-10)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>