progress.c: update progress every 50ms, not every 1 second
The progress bars git uses have updated every second since
96a02f8f6d2 (common progress display support, 2007-04-18). Let's
increase this by 20x to 50ms to make the UI appear more snappy, and
to (in subsequent commits) make "stalling" more apparent.

When running e.g. "git gc" the "Counting objects" number is updated
infrequently enough, while growing very fast, to appear to me to be
very slow flashes of new numbers, not something that's perceived to
"count up" quickly.

This change probably has a negative impact on cases where we're
sending the progress bar over the network, since we'll now have a lot
more network chatter. We should probably add an API to allow programs
to pass in a "my client is remote" or equivalent, to slow this update
rate down.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>