progress.[ch]: convert "title" to "struct strbuf"
Covert the "title" for the progress bar to a "struct strbuf", as with
the existing "counters_sb". Let's also rename the "counters_sb" to
merely "status", as we'll soon start using it not just to count, but
for any other arbitrary messaging after our fixed "title".

This makes the emitting the emitting code more consistent as it's now
consistently dealing with only a strbuf. Using the same pattern as
"status" also gives us a "title_len_utf8" variable. This should fix
any issues related to over-eager splitting of the progress bar in
languages where at typical character is not 1 byte.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>