pkt-line.[ch]: change packet_writer_int to *_INIT macro
Change the packet_writer_init() function to a corresponding macro,
this function didn't require anything we couldn't do in a simpler and
more idiomatic macro version.

This is a minor refactoring of code added in bc2e795cea6 (pkt-line:
introduce struct packet_writer, 2019-01-15), we'll make more use of
this patter in subsequent commits.

As this change shows nothing uses the fd != 1 API currently, so we
could remove that, but let's not in case we'll convert
e.g. get_remote_refs() in connect.c to this API in the future, it
requires specifying another FD.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>