ls-refs: move to "struct packet_writer" API
Move ls-refs to the "struct packet_writer" API added in
bc2e795cea6 (pkt-line: introduce struct packet_writer, 2019-01-15),
this makes use of the newly added PACKET_WRITER_INIT to simplify the
initialization, and do away with the effective hardcoding of "dest_fd
= 1" in two places.

Add and use a simple packet_writer_write_len() function as a
replacement for packet_write(). I think the end result is slightly
more pleasing to read, but this change is churny and really no real
for now, but in the next commit we'll make more use of the
packet_writer_*() API and make it worth it.

The API here is a bit of an existing mess, we have a packet_write()
which behaves like write(2), but had no corresponding
packet_writer_write(), instead that function took a format,
i.e. corresponded to packet_write_fmt().

We should really rename packet_writer_write() to
packet_writer_write_fmt(), and then rename this newly introduced
packet_writer_write_len() to packet_writer_write() for consistency.
But since we have a lot of existing packet_writer_write() users, let's
avoid that churn and live with the inconsistency for now.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>