upload-pack tests: test exact stdout/stderr on error
Amend the tests added in 9f9aa761305 (upload-pack: Improve error
message when bad ref requested, 2010-07-31) and
014ade74842 (upload-pack: send ERR packet for non-tip objects,
2019-04-13) to compare the exact error and protocol output we expect
on error.

I'm about to change the code that emits this output in a subsequent
commit. Let's assert that only the things we expect to change in the
output are going to change, as opposed to just grepping some
known-good and known-bad strings in the output.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>