protocol v2: correctly send ERR packet in "object-info"
Change the protocol v2 "object-info" to immediately error out on known
arguments. The behavior of the code added in a2ba162cda2 (object-info:
support for retrieving object info, 2021-04-20) goes against the
specification, i.e:

    Once [the ERR] packet is sent by a client or a server, the data
    transfer process defined in this protocol is terminated.

If it was sent lines that it would error on it would emit an ERR like
any other line response, and then continue serving requests.

Nothing in a2ba162cda2 suggests that this was intentional, the
behavior seems to have been a result of the ls-refs.c code having been
copied over and adjusted.

This makes use of the new packet_client_error() helper function, which
allows us to emit an ERR packet (like packet_writer_error()), and then
die().

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>