serve tests: move & improve "unexpected line" test
Improve the test added in 7cc6ed2d06d (upload-pack: fix error message
typo, 2018-05-01) to:

 1. Move it to the top, the comments in the file indicate that we have
    sections for ls-refs and object-info, this was in the ls-refs
    section. Let's add a comment to indicate this.

 2. We don't need to "git init" a repository for this, the test helper
    doesn't require it, let's not do it.

 3. Test both "tag" and "key=value" paramaters for completeness.

 4. Assert the full output instead of a "grep".

This better tests the "fetch" code added in 4c324c00501 (upload-pack:
Use finish_{command,async}() instead of waitpid()., 2007-11-04). As an
aside there are also tests for the behavior in
t/t5530-upload-pack-error.sh, but these are the raw protocol tests,
those are the integration tests.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>