protocol v2: test "fetch" error handling
Add real-world tests for how injecting a bad argument into the "fetch"
dialog in protocol v2 is handled by the server by using a
GIT_TEST_PROTOCOL_BAD_FETCH=true variable picked up by
fetch-pack.c. This compliments the existing tests for serve.c that
test the same behavior.

We can mostly test_cmp these, an exception is the strerror() case
noted, and that we might racily get:

    fatal: the remote end hung up unexpectedly

Because we sometimes die(), and depending on the client/server state
the client might emit that, or not. I'll fix up that behavior more
generically in follow-up work to make the server emit ERR lines and
cleanly disconnect in such cases, but let's grep these out for now.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>