protocol v2: send ERR on unknown in "fetch"
Change the protocol v2 "fetch" command to send an ERR packet when it
encounters unknown arguments, using the newly added
packet_client_error() API.

Because we simply died before the client would e.g. under HTTP see a
very cryptic error about not being able to read the "acknowledgments"
section header. As shown in the test output we now are able to point
to specific problems in the protocol dialog, and to have those
reported by clients.

There are still a lot of cases where upload-pack simply calls die()
for one thing or another, some of those will be migrated over to this
API in subsequent commits.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>