protocol v2 tests: test exact stdout/stderr on error
Amend the existing protocol v2 tests in t5702-protocol-v2.sh to
test_cmp the exact output on error, as opposed to just grepping
substrings of expected output.

This is more verbose, but gives us much better testing and
understanding of desired behavior. In some follow-up work I'll be
adjusting some of these error messages.

By using that "grep" pattern we hide potential changes, and make both
new features and changes to them harder to understand.

E.g. in 5476e1efded (fetch-pack: print and use dangling .gitmodules,
2021-02-22) we added the "disallowed submodule name" warning, the test
now makes it clear that we redundantly emit that exact warning twice.

I'll solve that bug separately, but I think it makes the point that we
should be using test_cmp. I'm pretty sure we'd have fixed it at the
time if we did that consistently, i.e. "why do we need the same
warning twice?" would have come up during review.

As noted in a preceding commit we are grepping out the "fatal: the
remote end hung up unexpectedly" messages (if any) for now. They'll be
addressed in follow-up work.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>