fsck: fix over-reporting of FSCK_MSG_GITMODULES_NAME
As shown in improved test added in earlier commits to the protocol v2
tests, we introduced the unwanted behavior in 5476e1efded (fetch-pack:
print and use dangling .gitmodules, 2021-02-22) of warning N times for
a bad "<name>" seen in a "submodule.<name>" section in a ".gitmodules"
files, where N is the number of config variables within that section.

This happened because in fsck_blob() we set up a "struct
fsck_gitmodules_data", end then proceed to call "git_config_from_mem"
with a "config_fn_t" callback. For a submodule section like:

    [submodule ".."]
    path = include/foo
    url = git://example.com/git/lib.git

We will be called twice, once for each of "path" and "url", if we had
a third variable we'd be called thrice, and so on.

Let's solve this by also using a "config_parser_event_fn_t"
callback. We'll get a CONFIG_EVENT_SECTION event as we enter a
section. We'll then only print the warning about the section name when
we on seeing see the first config after a given section.

The reason we don't have an "in_event" or "prev_event" to track this
is that multiple other events may occur between the section and the
first variable, e.g. multiple alternating CONFIG_EVENT_WHITESPACE and
CONFIG_EVENT_COMMENT.

There's no event corresponding to the "config_fn_t" callback gets
called for, it's an implicit "other". See the parsing & event
callbacks in git_parse_source(). The get_value() at the end of the
parsing loop is what calls the "config_fn_t" callback, before that we
handle any number of CONFIG_EVENT_*.

So let's juggle "first_after_section" and only warn on bad section
names when we see the first variable after a section name. We *do*
want to warn if there's multiple bad sections in the same .gitmodules
file, so it's not a boolean "have warned" for each
fsck_blob(). Furthermore, a repo may have multiple bad blobs with the
same bad path, so it should not be a global "have seen and warned
about this bad name" either.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>