fsck.c: clarify fsck_gitmodules_fn() by using "else if"
Change a chain of if/if/if added in 5476e1efded (fetch-pack: print and
use dangling .gitmodules, 2021-02-22) to use if/else if/else if
instead.

It's not possible for us to be in a "submodule.<name>.FOO" and
"submodule.<name>.BAR" section at the same time, that was clear from a
reading of the code and API, but let's further assure both the reader
and the compiler that we're only picking one of these branches.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>