fetch tests: fix cleanup logic in --negotiate-only tests
Fix an odd inter-dependency between the tests added in
9c1e657a8fd (fetch: teach independent negotiation (no packfile),
2021-05-04). We cleaned up at the start in "setup_negotiate_only", so
the wait-for-done was relying on not cleaning up the "server"
directory of the previous test, but was relying on cleaning up the
"client" directory (whose path doesn't change).

Let's factor these out into two helper functions, and add "&&" so
we'll spot failures within them. Subsequent commits will cleanup the
general logic in this test file.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>