protocol v2 tests: stress (often strange) packfile-uri behavior
As noted in a preceding commit the "disallowed submodule name" warning
is printed out twice or more in some cases. I'll address that in
follow-up work, but for now let's stress various edge cases related to
that.

Aside from that bug the first two tests here behave as I'd expect, but
there's some really weird behavior in the third one I haven't fully
dug into. I.e. when we fsck a repo we'll find and report issues with
.gitmodules files on all branches, with packfile-uri which issues we
note depend on what branches the parent repository has checked out.

I initially though I'd found a minor security issue with that,
i.e. you could put a bad .gitmodules file not-your-main branch, and
then trick someone to clone the repo with packfile-uri, if they used
the topic branch and "git submodule update --init" they could be
vulnerable.

But I'm fairly sure that's not the case, i.e. if we use the skipList
to exclude only those commits we'd get complaints about we'll find
that the fsck now picks up the previously missed commits, as the test
demonstrates. So it's some lazyness/early exit behavior in
index-pack. I haven't fully dug into it.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>