protocol v2: use helper for "ERR" on bad OID "list"
The series of "have"/"done" lines can be treated like a line where we
expected a "have" line for the purposes of this error message.

We could have a custom message here saying that we didn't get
something valid, but the wording of the
packet_client_error_expected_oid() error is general enough to work for
us, i.e.:

	static const char *msg = N_("%s: protocol error, "
				    "expected to get object ID on '%s' line, "
				    "not '%s'");

Perhaps we won't benefit from the same ambiguity in other languages,
but in practice if anyone sees this it's likely because it came over
with an ERR line, which won't be translated.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>