refs file-backend.c: stop setting "EBUSY" in verify_lock()
This EBUSY was set in 835e3c992fd (refs.c: verify_lock should set
errno to something meaningful, 2014-06-20).

That was done to fix a bug in code that was later refactored in
cd94f765720 (fetch.c: change s_update_ref to use a ref transaction,
2014-04-28) to not use errno at all.

Those two changes were were a part of the same series, but the end
result is that we ended up with this dead code. Since the "save_errno"
here was purely in service of the "errno = EBUSY" we can remove it
too.

Removing this makes subsequent commits where we refactor code
surrounding verify_lock() (which may use errno) easier to reason
about.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>