reflog expire: pass OID via cb_data, not reflog_expire()
Pass the OID to our own callbacks via our own "cb_data", instead
relying on the "oid" parameter that the refs.c API will provide our
reflog_expiry_prepare() with. The OID that API passed us was just our
own unmodified "OID", as the added assertion confirms.

This doesn't change any logic, but makes a a subsequent change where
we remove the "oid" parameter from the refs.[ch] API smaller, this was
its last user.

The "oid" parameter to dwim_log() is now unused, it would be easy to
change it to support NULL. It calls refs_resolve_ref_unsafe() with our
supplied OID, which already supports that, but then does its own
oidcpy() to our target. Let's just change the name of our variable to
"unused" for now. That dwim_log() is being done to get the name into
the "ref" variable, not to get the OID.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>